contracts - adapters v1

+++ Modules/Product/app/Services/ProductService.php
@@ -5,6 +5,8 @@ namespace Modules\Product\Services;
 use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Collection;
 use Modules\Category\Models\Category;
+use Modules\Product\Contracts\Review\ContractsReview;
+use Modules\Product\Exceptions\ProductHasReviewsException;
 use Modules\Product\Models\Product;
 use Modules\Product\Repositories\ProductRepositoryInterface;
 
@@ -12,6 +14,7 @@ class ProductService implements ProductServiceInterface
 {
     public function __construct(
         private readonly ProductRepositoryInterface $products,
+        private readonly ContractsReview $reviews,
     ) {}
 
     public function getAllProducts(): Collection
@@ -27,6 +30,11 @@ class ProductService implements ProductServiceInterface
     public function getProductById(int $id): Product
     {
         return $this->products->findOrFail($id);
+    }
+
+    public function getReviewsByProductId(int $productId): Collection
+    {
+        return $this->reviews->getByProductId($productId);
     }
 
     public function getCategoryOptions(): Collection
@@ -52,6 +60,10 @@ class ProductService implements ProductServiceInterface
     public function deleteProduct(int $id): ?bool
     {
         $product = $this->products->findOrFail($id);
+
+        if ($this->reviews->existsInProduct($product->id)) {
+            throw new ProductHasReviewsException('Cannot delete product because it has reviews.');
+        }
 
         return $this->products->delete($product);
     }
